* Multirate real file support
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@671 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/lib/video_win32.c
@@ -325,7 +325,7 @@ static int init_std(bgav_win32_thread_t * thread)
   pack_bih(&priv->bih_in, &bih_in);
 
   fprintf(stderr, "ICDecompressGetFormatSize: %d\n",
-          ICDecompressGetFormatSize(priv->hic, &bih_in));
+          (int)ICDecompressGetFormatSize(priv->hic, &bih_in));
           
   result = ICDecompressGetFormat(priv->hic, &priv->bih_in, &priv->bih_out);
